The NTLM Challenge message's targetNameBufferOffset is 32bit (not 16)
This doesn't fix any issues that I'm aware of since the 16bit value
we were reading before is enough to store the full value (the other 2
bytes are always 0's).

+++ MailKit/Security/Ntlm/NtlmChallengeMessage.cs
@@ -98,7 +98,8 @@ namespace MailKit.Security.Ntlm {
 			Buffer.BlockCopy (message, startIndex + 24, serverChallenge, 0, 8);
 
 			var targetNameLength = BitConverterLE.ToUInt16 (message, startIndex + 12);
-			var targetNameOffset = BitConverterLE.ToUInt16 (message, startIndex + 16);
+			//var targetNameMaxLength = BitConverterLE.ToUInt16 (message, startIndex + 14);
+			var targetNameOffset = BitConverterLE.ToInt32 (message, startIndex + 16);
 
 			if (targetNameLength > 0) {
 				var encoding = (Flags & NtlmFlags.NegotiateUnicode) != 0 ? Encoding.Unicode : Encoding.UTF8;
